requirements.txt
blinker==1.6.2
click==8.1.7
dnspython==2.4.2
email-validator==2.0.0.post2
Flask==2.3.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
idna==3.4
infinity==1.5
intervals==0.9.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Pillow==10.1.0
Pympler==1.0.1
six==1.16.0
validators==0.22.0
Werkzeug==2.3.7
WTForms==3.0.1
WTForms-Components==0.10.5
